rename course progress middleware
